requirements/prod.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt
psycopg2-binary
Pillow